Added adapters, addons, and the context.
Adapters implement the "generic function" pattern.  The interface is inspired
by Zope Component Architecture, although the semantics of the multiple
displatch is borrowed from Cecil.

Addons provide a way to extend an HTSQL application with extra adapters.

The context is a thread-local variable that keeps the active application.
Since there could be more than one HTSQL application in the same
Python process, and different application may use different sets of
adapters, we need to either pass the current application to every API
function or keep the current application in a global variable.