Change format for data references.
Now data reference can be either a string (URL) or a mapping with "url" and
"refs" keys. Refs are not manually specified instead of parsing them from query
string.

DataReference is now an abstract base class. Its former functionality now
resides in CollectionReference. Implementations for PaginatedCollectionReference
and EntityRefernce are pending.