DataPreloader: component which automates <Preloader /> rendering for data components.
Example:

    <DataPreloader data={this.props.data}>
      <Grid data={this.props.data} />
    </DataPreloader>

Will only render `<Grid />` if `this.props.data` is not in `updating` state,
otherwise it renders `<Preloader />`.

Another use case is:

    <DataPreloader
      data={this.props.data}
      renderer={this.renderGrid}
      />

Which is similar to the first one but will call a function `renderer` in case
data is available. This is useful when you want to render all the data inside a
single component.