Fix modal styles to reset white-space CSS prop
Header overrides `white-space` with `nowrap` for its items which then appears to
have effect on modal windows which were initiated from the header elements.

The right fix would be to open modal by attaching its DOM node to `<body>`
directly so those CSS props are not inherited.

This is a quickfix instead (but working and generic for each modal) which just
resets `white-space` CSS prop back to `normal` for modal content.

Fixes PLATFORM-1748