rex.query: add query variables
This adds "global" query variables to rex.query.

`RexBindingState` now can be initialized with a mapping from names to variables
values. Variable values are expected to be just python values which will be
treated as literals or pre-made bindings.

The last bit is needed for rex.graphql as we use this to populate mapping with
bindings which represent typecases like `(null : type)` so we can validate query
structure and infer types without knowing the concrete argument values. This is
not super elegant so I'd welcome suggestons here very much.

Then the variables are represented as `app(var, list(val))` in syntax and being
translated as lookups into the mapping described above.

Some more notes:

- Variables cannot be used with take - `app(take, var(name))` is invalid. This
  is because variable values can be represented as bindings and HTSQL model
  doesn't allow that.

- Variables cannot be lists. This would be useful for rex.graphql though as it
  would allow `q.id == argument("id", List(scalar.ID))` queries.