htsql: fix record domain basis by including field tags
This fixes a bug with generation of record classes for queries.

For example, previously the following queries generated the same plan shared
between each other (it is cached):

    /region {name}
    /region {region_name := name}

which means that record class was shared between those queries. This is not
correct as the order of which query of those executes first will matter for
record class fields — it will have either `name` or `region_name` attribute.