rex.graphql: add opaque types
This commit introduces opaque types to rex.graphql. It is a way to
define new scalars which wraps existing scalar but give them a new name.

That's useful as those scalars won't be unified with other scalars and
therefore you cannot mix them in arg lists. Also for rex-graphql-codegen
this will allow to generate more precise types.