js: update flow to 0.118.0
This causes some new errors in code. I suspect some just require
different (more) annotations but other are actual bugs. This is mostly
old code though so for now in this PR I just suppressed errors so we
don't change the behaviour and cause new bugs.

Once we merge this we need to propagate this to other codebases whete we
can see more errors appear too.