rex.setup: pin docutils to >=0.12, <0.13
This is the version pinned in rex.widget and rex.action so this is
what's used in the codebase.

But in the rex.setup's tests we are installing only rex.setup and
rex.setup_demo and the newly released docutils 0.18 breaks the old
Sphinx we are using. So the pin of docutils version at rex.setup fixes
this.